pyproject.toml
[tool.poetry]
name = "architex"
version = "0.1.8"
description = "Draw your software architecture diagram automagically"
authors = ["Miko <jherjati@gmail.com>"]
readme = "README.md"
packages = [{include = "architex"}]

[tool.poetry.dependencies]
python = "^3.7"
pyyaml = "^5.0"
diagrams = "^0.23.3"
crossplane = "^0.5.8"
typer = "0.3.2"
colorama = "0.4.4"
shellingham = "1.4.0"
pytest = "6.2.4"


[tool.poetry.group.dev.dependencies]
autopep8 = "^2.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
